requirements.txt
fastapi
uvicorn
requests
python-dotenv
PyYAML
gitpython